% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPathwaySeparation.R
\name{getAssociatedRAVs}
\alias{getAssociatedRAVs}
\title{check whether your pathway of interest is in the model}
\usage{
getAssociatedRAVs(PCAmodel, pathwaySet, cutoff_nes, cutoff_n)
}
\arguments{
\item{PCAmodel}{A PCAGenomicSignatures object.}

\item{pathwaySet}{A character vector. Keywords you are looking for.}

\item{cutoff_nes}{A minimum NES to be kept.}

\item{cutoff_n}{A number of pathways to keep.}
}
\description{
check whether your pathway of interest is in the model
}
