% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPathwaySeparation.R
\name{checkPathwaySeparation}
\alias{checkPathwaySeparation}
\title{Check a pair of pathways are captured in the signature model}
\usage{
checkPathwaySeparation(
  PCAmodel,
  pathway1,
  pathway2,
  cutoff_nes = NULL,
  cutoff_n = 5
)
}
\arguments{
\item{PCAmodel}{PCAGenomicSignatures object.}

\item{cutoff_nes}{A minimum NES to be kept. Under the default (\code{NULL}), this
filter is not used.}

\item{pathway}{A character vector. Keywords you are looking for.}

\item{cutoof_n}{A number of pathways to keep.Default is 5. If you don't want to
use this filter, assign it as \code{NULL}.}
}
\description{
Check a pair of pathways are captured in the signature model
}
\note{
https://github.com/greenelab/multi-plier/blob/master/32-explore_pathway_separation.Rmd
}
