% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVScatter.R
\name{LVScatter}
\alias{LVScatter}
\title{Draw scatter plot with R-squared value}
\usage{
LVScatter(dat, lv, y.var = "Neutrophil.Count", ylab = "")
}
\arguments{
\item{dat}{A data frame.}

\item{lv}{A character.}

\item{y.var}{}
}
\value{
A scatter plot.
}
\description{
Draw scatter plot with R-squared value
}
\note{
Modified from https://github.com/greenelab/multi-plier/blob/master/07-sle_cell_type_recount2_model.Rmd
}
